all trigonometry 100% coverage

+++ src/s21_fabs.c
@@ -1,5 +1,5 @@
 #include "s21_math.h"
 
-double s21_fabs(double x) {
+long double s21_fabs(double x) {
   return x < 0 ? x * -1 : x;
 }
